G120: Compile

+++ Drivers/AT45DB321D_Flash/AT45DB321D_Flash.h
@@ -9,4 +9,4 @@ TinyCLR_Result AT45DB321D_Flash_Write(uint32_t address, size_t length, const uin
 TinyCLR_Result AT45DB321D_Flash_EraseBlock(uint32_t sector);
 TinyCLR_Result AT45DB321D_Flash_IsBlockErased(uint32_t sector, bool &erased);
 TinyCLR_Result AT45DB321D_Flash_GetBytesPerSector(uint32_t address, int32_t& size);
-TinyCLR_Result AT45DB321D_Flash_GetSectorMap(const uint32_t*& addresses, const uint32_t*& sizes, size_t& count);
+TinyCLR_Result AT45DB321D_Flash_GetSectorMap(const uint64_t*& addresses, const size_t*& sizes, size_t& count);
